+++ jkube-kit/enricher/generic/src/test/java/org/eclipse/jkube/enricher/generic/ingress/NetworkingV1IngressGeneratorTest.java
@@ -19,6 +19,7 @@ import io.fabric8.kubernetes.api.model.ServiceBuilder;
 import io.fabric8.kubernetes.api.model.ServiceSpecBuilder;
 import io.fabric8.kubernetes.api.model.networking.v1.Ingress;
 import io.fabric8.kubernetes.api.model.networking.v1.IngressSpec;
+import org.assertj.core.api.InstanceOfAssertFactories;
 import org.eclipse.jkube.kit.config.resource.IngressRuleConfig;
 import org.eclipse.jkube.kit.config.resource.IngressRulePathConfig;
 import org.eclipse.jkube.kit.config.resource.IngressRulePathResourceConfig;
@@ -37,7 +38,7 @@ class NetworkingV1IngressGeneratorTest {
         ServiceBuilder testSvcBuilder = initTestService();
 
         // When
-        Ingress ingress = NetworkingV1IngressGenerator.generate(testSvcBuilder, "org.eclipse.jkube", null, Collections.emptyList(), Collections.emptyList());
+        Ingress ingress = NetworkingV1IngressGenerator.generate(testSvcBuilder, "org.eclipse.jkube", null, null, Collections.emptyList(), Collections.emptyList());
 
         // Then
         assertThat(ingress).isNotNull()
@@ -56,7 +57,7 @@ class NetworkingV1IngressGeneratorTest {
         ServiceBuilder testSvcBuilder = initTestService();
 
         // When
-        Ingress ingress = NetworkingV1IngressGenerator.generate(testSvcBuilder, null, null, Collections.emptyList(), Collections.emptyList());
+        Ingress ingress = NetworkingV1IngressGenerator.generate(testSvcBuilder, null, null, null, Collections.emptyList(), Collections.emptyList());
 
         // Then
         assertThat(ingress).isNotNull()
@@ -90,12 +91,54 @@ class NetworkingV1IngressGeneratorTest {
                 .build();
 
         // When
-        Ingress ingress = NetworkingV1IngressGenerator.generate(testSvcBuilder, "org.eclipse.jkube", null, Collections.singletonList(ingressRuleConfig), Collections.singletonList(ingressTlsConfig));
+        Ingress ingress = NetworkingV1IngressGenerator.generate(testSvcBuilder, "org.eclipse.jkube", null, null, Collections.singletonList(ingressRuleConfig), Collections.singletonList(ingressTlsConfig));
 
         // Then
         assertThat(ingress)
             .hasFieldOrPropertyWithValue("metadata.name", "test-svc")
             .extracting("spec.rules").asList()
+            .singleElement()
+            .hasFieldOrPropertyWithValue("host", "foo.bar.com");
+    }
+
+    @Test
+    void generate_withIngressClassName() {
+        // Given
+        ServiceBuilder testSvcBuilder = initTestService();
+
+        // When
+        Ingress ingress = NetworkingV1IngressGenerator.generate(testSvcBuilder, "org.eclipse.jkube", null, "nginx", Collections.emptyList(), Collections.emptyList());
+
+        // Then
+        assertThat(ingress).isNotNull()
+            .hasFieldOrPropertyWithValue("metadata.name", "test-svc")
+            .extracting("spec").isNotNull()
+            .hasFieldOrPropertyWithValue("ingressClassName", "nginx");
+    }
+
+    @Test
+    void generate_withIngressClassNameAndXMLConfig() {
+        // Given
+        ServiceBuilder testSvcBuilder = initTestService();
+        IngressRuleConfig ingressRuleConfig = IngressRuleConfig.builder()
+                .host("foo.bar.com")
+                .path(IngressRulePathConfig.builder()
+                        .path("/foo")
+                        .pathType("Prefix")
+                        .serviceName("test-svc")
+                        .servicePort(8080)
+                        .build())
+                .build();
+
+        // When
+        Ingress ingress = NetworkingV1IngressGenerator.generate(testSvcBuilder, "org.eclipse.jkube", null, "traefik", Collections.singletonList(ingressRuleConfig), Collections.emptyList());
+
+        // Then
+        assertThat(ingress).isNotNull()
+            .hasFieldOrPropertyWithValue("metadata.name", "test-svc")
+            .extracting("spec").isNotNull()
+            .hasFieldOrPropertyWithValue("ingressClassName", "traefik")
+            .extracting("rules").asInstanceOf(InstanceOfAssertFactories.LIST)
             .singleElement()
             .hasFieldOrPropertyWithValue("host", "foo.bar.com");
     }